+++ country.go
@@ -70,7 +70,7 @@ func readCountryMap(country *Country, buffer []byte, mapSize uint, offset uint) 
 				return 0, err
 			}
 		case "confidence":
-			country.Confidence, offset, err = readUInt8(buffer, offset)
+			country.Confidence, offset, err = readUInt16(buffer, offset)
 			if err != nil {
 				return 0, err
 			}
